// packages/mpl-core-kit-lib/src/generated/types/baseLinkedLifecycleHook.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  type Address,
  type Codec,
  combineCodec,
  type Decoder,
  type Encoder,
  getAddressDecoder,
  getAddressEncoder,
  getArrayDecoder,
  getArrayEncoder,
  getOptionDecoder,
  getOptionEncoder,
  getStructDecoder,
  getStructEncoder,
  type Option,
  type OptionOrNullable,
} from '@solana/kit'

import {
  type BaseExtraAccount,
  type BaseExtraAccountArgs,
  type BasePluginAuthority,
  type BasePluginAuthorityArgs,
  type ExternalPluginAdapterSchema,
  type ExternalPluginAdapterSchemaArgs,
  getBaseExtraAccountDecoder,
  getBaseExtraAccountEncoder,
  getBasePluginAuthorityDecoder,
  getBasePluginAuthorityEncoder,
  getExternalPluginAdapterSchemaDecoder,
  getExternalPluginAdapterSchemaEncoder,
} from './index.ts'

export type BaseLinkedLifecycleHook = {
  dataAuthority: Option<BasePluginAuthority>
  extraAccounts: Option<Array<BaseExtraAccount>>
  hookedProgram: Address
  schema: ExternalPluginAdapterSchema
}

export type BaseLinkedLifecycleHookArgs = {
  dataAuthority: OptionOrNullable<BasePluginAuthorityArgs>
  extraAccounts: OptionOrNullable<Array<BaseExtraAccountArgs>>
  hookedProgram: Address
  schema: ExternalPluginAdapterSchemaArgs
}

export function getBaseLinkedLifecycleHookCodec(): Codec<BaseLinkedLifecycleHookArgs, BaseLinkedLifecycleHook> {
  return combineCodec(getBaseLinkedLifecycleHookEncoder(), getBaseLinkedLifecycleHookDecoder())
}

export function getBaseLinkedLifecycleHookDecoder(): Decoder<BaseLinkedLifecycleHook> {
  return getStructDecoder([
    ['hookedProgram', getAddressDecoder()],
    ['extraAccounts', getOptionDecoder(getArrayDecoder(getBaseExtraAccountDecoder()))],
    ['dataAuthority', getOptionDecoder(getBasePluginAuthorityDecoder())],
    ['schema', getExternalPluginAdapterSchemaDecoder()],
  ])
}

export function getBaseLinkedLifecycleHookEncoder(): Encoder<BaseLinkedLifecycleHookArgs> {
  return getStructEncoder([
    ['hookedProgram', getAddressEncoder()],
    ['extraAccounts', getOptionEncoder(getArrayEncoder(getBaseExtraAccountEncoder()))],
    ['dataAuthority', getOptionEncoder(getBasePluginAuthorityEncoder())],
    ['schema', getExternalPluginAdapterSchemaEncoder()],
  ])
}